# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: NAPTIN Enterprise Portal API
  description: |
    Contract sketch for a future backend aligned with portal modules.
    All paths are prefixed by `/api/v1` unless behind a gateway.
    Authentication — Bearer JWT; roles — staff | hod | director | ict_admin | auditor.
  version: 0.1.0
servers:
  - url: https://portal.naptin.gov.ng/api/v1
    description: Production (illustrative)
  - url: http://localhost:8080/api/v1
    description: Local dev
...
paths:
...
  /workbench/summary:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Top-level workbench metrics for Client Ops and New Markets
  /workbench/onboarding:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: List onboarding records
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Create onboarding workflow from closed-won trigger
...
  /workbench/opportunities:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Upsell and cross-sell opportunities
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Create opportunity
  /workbench/markets/candidates:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Candidate markets with weighted scores
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Add candidate market
  /workbench/pilots:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Market entry pilot records
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Create pilot charter
  /workbench/feedback:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Client feedback queue and routing status
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Submit feedback item
  /workbench/renewals:
    get:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Contract renewal pipeline
    post:
      tags: [client_ops_new_markets]
      security: [{ bearerAuth: [] }]
      summary: Create renewal proposal
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...